refactor: modularity split and console scrollback fix
