+++ cli/tracking.go
@@ -2,13 +2,11 @@ package cli
 
 import (
 	"fmt"
-	"focusd/core"
+	"focusd/coreapi"
 	"focusd/storage"
 	"focusd/system"
 	"focusd/ui"
-	"log"
 	"os"
-	"time"
 )
 
 func runStart() {
@@ -47,7 +45,7 @@ func runStop() {
 		return
 	}
 
-	if core.SendIPCCmd("stop") {
+	if coreapi.SendIPCCmd("stop") {
 		ui.PrintOK("focusd stopped gracefully")
 		return
 	}
@@ -56,28 +54,5 @@ func runStop() {
 		ui.PrintWarn("Could not stop focusd. It may still be running.")
 	} else {
 		ui.PrintOK("focusd stopped (forced)")
-	}
-}
-
-func RunDaemon() {
-	var dbErr error
-	for i := 0; i < 5; i++ {
-		dbErr = storage.Init()
-		if dbErr == nil {
-			break
-		}
-		time.Sleep(1 * time.Second)
-	}
-	if dbErr != nil {
-		os.Exit(1)
-	}
-	defer storage.Close()
-
-	storage.EnforceRetention()
-
-	tracker := core.NewTracker()
-	if err := tracker.Start(); err != nil {
-		log.Printf("ERROR: daemon failed to start tracker: %v", err)
-		os.Exit(1)
 	}
 }
